kill `server_process` more smartly
Tests that spawns `server_process` are supposed to kill the process
before it exits. Otherwise, other tests that also need server will not
run.

Previously, it naively did `finally: server_process.kill()`. It worked
because any exceptions will be caught later. But it hides the *real*
exception and throws "server_process is not assigned". The tester has no
idea why the test failed.

+++ tests/clone.py
@@ -20,6 +20,7 @@ from utils import (
 
 def clone(base2_size: int = 600):
     goto_root()
+    server_process = None
 
     try:
         server_process = spawn_ragit_server()
@@ -113,4 +114,5 @@ def clone(base2_size: int = 600):
             os.chdir("..")
 
     finally:
-        server_process.kill()
+        if server_process is not None:
+            server_process.kill()
